+++ src/extension.ts
@@ -6,8 +6,18 @@ import * as utils from './utils';
 
 const decorRanges: utils.DecorRange[] = [];
 const documentsContent: utils.DocumentContent[] = [];
+const rulerPosition = {
+    Center: 2,
+    Left: 1,
+    Right: 4,
+    Full: 7
+}
 
 export async function activate(context) {
+    if(await utils.checkForGitRepo(context)){
+        return(null);
+    }
+
     utils.readConfig();
     // await utils.checkForGitPresence(context);
     utils.checkForOutputOption(context);
@@ -142,7 +152,7 @@ function createDecorator(type: string): vscode.TextEditorDecorationType {
     if (utils.config.showInOverView) {
         obj = Object.assign(obj, {
             overviewRulerColor: hexToRgba(utils.overviewConfig[type], utils.overviewConfig.opacity),
-            overviewRulerLane: 2,
+            overviewRulerLane: rulerPosition[utils.overviewConfig.position],
         });
     }